ski-bootloader: Increase frequency for SAL_FREQ_BASE_PLATFORM
This simple change yields between 1.19 and 1.53 times as much performance
for writes to ext4 file systems inside ski compared to the stock value.

For package building this yields - depending on the task (`configure`,
`make`, `make install`, `makepkg`) and package - between 1.25 and 1.47
times as much performance and between 1.32 and 1.36 times the performance
overall.

Another effect of this change is that the clock inside ski runs much slower
and closer to the host's clock.

Tested on a machine with 1 x Xeon E3 1271v3 with a maximum clock of 4 GHz.
FS images for ski were placed on a SATA SSD on a 6 Gbps link. Package
builds happened in tmpfs, so in memory inside ski. Timings for the
performed tests were gathered on the host.

The used kernel was based on:
* https://github.com/johnny-mnemonic/linux-ia64/tree/v6.13-rc1-w-ia64
* and https://github.com/linux-ia64/ski/blob/hp-sim-for-linux/linux-6.12.y/linux-6.12.0-reinstate-hp-sim-platform.patch
...on top and used the stock `arch/ia64/configs/sim_defconfig` with tmpfs enabled.